# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedTrack Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: patientId
          schema:
            type: integer
        - in: query
          name: doctorId
          schema:
            type: integer
        - in: query
          name: appointmentId
          schema:
            type: integer
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create a prescription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionRequest"
      responses:
        "201":
          description: Prescription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get prescription by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prescription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Add a staff member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Prescription:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        diagnosis:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrescriptionItem"
        createdAt:
          type: string
      required:
        - id
        - patientId
        - doctorId
        - diagnosis
        - items
        - createdAt
    PrescriptionItem:
      type: object
      properties:
        id:
          type: integer
        medicineName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: ["string", "null"]
      required:
        - id
        - medicineName
        - dosage
        - frequency
        - duration
    CreatePrescriptionRequest:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        diagnosis:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreatePrescriptionItemRequest"
      required:
        - patientId
        - doctorId
        - diagnosis
        - items
    CreatePrescriptionItemRequest:
      type: object
      properties:
        medicineName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: ["string", "null"]
      required:
        - medicineName
        - dosage
        - frequency
        - duration
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        department:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        username:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - role
        - username
        - createdAt
    CreateStaffRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        department:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        username:
          type: string
        password:
          type: string
      required:
        - name
        - role
        - username
        - password
...