# lib/api-zod/src/generated/types/createPrescriptionItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedTrack Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePrescriptionItemRequest {
  medicineName: string;
  dosage: string;
  frequency: string;
  duration: string;
  /** @nullable */
  instructions?: string | null;
}