# lib/api-zod/src/generated/types/createStaffRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedTrack Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStaffRequest {
  name: string;
  role: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  username: string;
  password: string;
}